+++ pkg/helm/client.go
@@ -1,17 +1,13 @@
 package helm
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/containerd/containerd/log"
-	"groundcover.com/pkg/k8s"
 	"groundcover.com/pkg/utils"
 	"helm.sh/helm/v3/pkg/action"
 	"helm.sh/helm/v3/pkg/cli"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
-	"k8s.io/client-go/rest"
 )
 
 type Client struct {
@@ -37,17 +33,7 @@ func NewHelmClient(namespace, kubecontext string) (*Client, error) {
 	helmClient.settings.SetNamespace(namespace)
 	helmClient.settings.KubeContext = kubecontext
 
-	getter, ok := helmClient.settings.RESTClientGetter().(*genericclioptions.ConfigFlags)
-	if !ok {
-		return nil, fmt.Errorf("failed to cast helm rest client getter")
-	}
-
-	getter.WrapConfigFn = func(restConfig *rest.Config) *rest.Config {
-		k8s.OverrideDepartedAuthenticationApiVersion(restConfig)
-		return restConfig
-	}
-
-	if err = helmClient.cfg.Init(getter, helmClient.settings.Namespace(), os.Getenv("HELM_DRIVER"), log.L.Debugf); err != nil {
+	if err = helmClient.cfg.Init(helmClient.settings.RESTClientGetter(), helmClient.settings.Namespace(), os.Getenv("HELM_DRIVER"), log.L.Debugf); err != nil {
 		return nil, err
 	}